+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+* `check` ignores config file unless explicitly given by commandline #2
+
 ## 0.2.0 (2017-12-26)
 
 * Add `version` command